updated number type

+++ storage.go
@@ -10,7 +10,7 @@ type Storage interface {
 	DeleteAccount(int) error
 	UpdateAccount(*TransferRequest) error
 	GetAccounts() ([]*Account, error)
-	GetAccountByNumber(int) (*Account, error)
+	GetAccountByNumber(int64) (*Account, error)
 	GetAccountByID(int) (*Account, error)
 }
